Move generated validators to src/ to simplify set-up and validate more types
