+++ .eslintrc
@@ -11,6 +11,7 @@
     "max-len": 0,
     "jsdoc/check-tag-names": [
       "error", {"definedTags": ["internal"]}
-    ]
+    ],
+    "filenames/match-exported": 0
   }
 }
