+++ .gitignore
@@ -2,6 +2,7 @@ node_modules
 package-lock.json
 coverage
 dist
+src/generated
 tsconfig.tsbuildinfo
 *.log
 .*
