+++ tsconfig.json
@@ -9,7 +9,7 @@
     "target": "ES2018",
     "module": "commonjs",
     "lib": ["ES2018"],
-    "allowJs": false,
+    "allowJs": true,
     "declaration": true,
     "sourceMap": true,
     "outDir": "dist",
@@ -22,7 +22,7 @@
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true,
     "moduleResolution": "node",
-    "typeRoots": ["types", "node_modules/@types"],
+    "typeRoots": ["src/types/external", "node_modules/@types"],
     "esModuleInterop": true,
     "stripInternal": true,
     "skipLibCheck": false,
